separate tests to C# 8 and before C# 7
